+++ bottom-bar/src/androidTest/java/com/roughike/bottombar/TabParserTest.java
@@ -27,7 +27,7 @@ public class TabParserTest {
     @Before
     public void setUp() throws Exception {
         context = InstrumentationRegistry.getContext();
-        tabs = new TabParser(context, new TabParser.Config.Builder().build(), com.example.bottombar.sample.test.R.xml.dummy_tab_xml)
+        tabs = new TabParser(context, new TabParser.Config.Builder().build(), com.roughike.bottombar.test.R.xml.dummy_tabs_five)
                 .getTabs();
     }
 
@@ -68,7 +68,7 @@ public class TabParserTest {
         assertEquals(Color.parseColor("#FF0000"), tabs.get(0).getActiveColor());
 
         assertEquals(
-                ContextCompat.getColor(context, com.example.bottombar.sample.test.R.color.test_random_color),
+                ContextCompat.getColor(context, com.roughike.bottombar.test.R.color.test_random_color),
                 tabs.get(1).getActiveColor()
         );
 
@@ -88,7 +88,7 @@ public class TabParserTest {
 
     @Test
     public void iconResourceIdsAsExpected() {
-        int expectedId = com.example.bottombar.sample.test.R.drawable.empty_icon;
+        int expectedId = com.roughike.bottombar.test.R.drawable.empty_icon;
 
         assertEquals(expectedId, tabs.get(0).getIconResId());
         assertEquals(expectedId, tabs.get(1).getIconResId());