+++ bottom-bar/src/androidTest/java/com/roughike/bottombar/ThreeFixedTabsTest.java
@@ -1,16 +1,12 @@
 package com.roughike.bottombar;
 
 import android.os.Bundle;
+import android.support.test.InstrumentationRegistry;
 import android.support.test.annotation.UiThreadTest;
 import android.support.test.filters.LargeTest;
-import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
-import com.example.bottombar.sample.R;
-import com.example.bottombar.sample.ThreeTabsActivity;
-
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InOrder;
@@ -30,13 +26,8 @@ import static org.mockito.Mockito.verifyZeroInteractions;
  */
 @RunWith(AndroidJUnit4.class)
 @LargeTest
-public class ThreeTabsActivityTest {
-    @Rule
-    public ActivityTestRule<ThreeTabsActivity> threeTabsActivityRule =
-            new ActivityTestRule<>(ThreeTabsActivity.class);
-
+public class ThreeFixedTabsTest {
     private OnTabSelectListener selectListener;
-
     private OnTabReselectListener reselectListener;
 
     private BottomBar bottomBar;
@@ -46,7 +37,8 @@ public class ThreeTabsActivityTest {
         selectListener = Mockito.mock(OnTabSelectListener.class);
         reselectListener = Mockito.mock(OnTabReselectListener.class);
 
-        bottomBar = (BottomBar) threeTabsActivityRule.getActivity().findViewById(R.id.bottomBar);
+        bottomBar = new BottomBar(InstrumentationRegistry.getContext());
+        bottomBar.setItems(com.roughike.bottombar.test.R.xml.dummy_tabs_three);
         bottomBar.setOnTabSelectListener(selectListener);
         bottomBar.setOnTabReselectListener(reselectListener);
     }
@@ -60,50 +52,50 @@ public class ThreeTabsActivityTest {
     @Test
     @UiThreadTest
     public void findingPositionForTabs_ReturnsCorrectPositions() {
-        assertEquals(0, bottomBar.findPositionForTabWithId(R.id.tab_favorites));
-        assertEquals(1, bottomBar.findPositionForTabWithId(R.id.tab_nearby));
-        assertEquals(2, bottomBar.findPositionForTabWithId(R.id.tab_friends));
+        assertEquals(0, bottomBar.findPositionForTabWithId(com.roughike.bottombar.test.R.id.tab_favorites));
+        assertEquals(1, bottomBar.findPositionForTabWithId(com.roughike.bottombar.test.R.id.tab_nearby));
+        assertEquals(2, bottomBar.findPositionForTabWithId(com.roughike.bottombar.test.R.id.tab_friends));
     }
 
     @Test
     @UiThreadTest
     public void whenTabIsSelected_SelectionListenerIsFired() {
-        bottomBar.selectTabWithId(R.id.tab_friends);
-        bottomBar.selectTabWithId(R.id.tab_nearby);
-        bottomBar.selectTabWithId(R.id.tab_favorites);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_friends);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_nearby);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_favorites);
 
         InOrder inOrder = inOrder(selectListener);
-        inOrder.verify(selectListener, times(1)).onTabSelected(R.id.tab_friends);
-        inOrder.verify(selectListener, times(1)).onTabSelected(R.id.tab_nearby);
-        inOrder.verify(selectListener, times(1)).onTabSelected(R.id.tab_favorites);
+        inOrder.verify(selectListener, times(1)).onTabSelected(com.roughike.bottombar.test.R.id.tab_friends);
+        inOrder.verify(selectListener, times(1)).onTabSelected(com.roughike.bottombar.test.R.id.tab_nearby);
+        inOrder.verify(selectListener, times(1)).onTabSelected(com.roughike.bottombar.test.R.id.tab_favorites);
         inOrder.verifyNoMoreInteractions();
     }
 
     @Test
     @UiThreadTest
     public void afterConfigurationChanged_SavedStateRestored_AndSelectedTabPersists() {
-        bottomBar.selectTabWithId(R.id.tab_favorites);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_favorites);
 
         Bundle savedState = bottomBar.saveState();
-        bottomBar.selectTabWithId(R.id.tab_nearby);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_nearby);
         bottomBar.restoreState(savedState);
 
-        assertEquals(R.id.tab_favorites, bottomBar.getCurrentTabId());
+        assertEquals(com.roughike.bottombar.test.R.id.tab_favorites, bottomBar.getCurrentTabId());
     }
 
     @Test
     @UiThreadTest
     public void whenTabIsReselected_ReselectionListenerIsFired() {
-        int firstTabId = R.id.tab_favorites;
+        int firstTabId = com.roughike.bottombar.test.R.id.tab_favorites;
         bottomBar.selectTabWithId(firstTabId);
         verify(reselectListener, times(1)).onTabReSelected(firstTabId);
 
-        int secondTabId = R.id.tab_nearby;
+        int secondTabId = com.roughike.bottombar.test.R.id.tab_nearby;
         bottomBar.selectTabWithId(secondTabId);
         bottomBar.selectTabWithId(secondTabId);
         verify(reselectListener, times(1)).onTabReSelected(secondTabId);
 
-        int thirdTabId = R.id.tab_friends;
+        int thirdTabId = com.roughike.bottombar.test.R.id.tab_friends;
         bottomBar.selectTabWithId(thirdTabId);
         bottomBar.selectTabWithId(thirdTabId);
         verify(reselectListener, times(1)).onTabReSelected(thirdTabId);
@@ -112,7 +104,7 @@ public class ThreeTabsActivityTest {
     @Test
     @UiThreadTest
     public void whenDefaultTabIsSet_ItsSelectedAtFirst() {
-        int defaultTabId = R.id.tab_friends;
+        int defaultTabId = com.roughike.bottombar.test.R.id.tab_friends;
 
         bottomBar.setDefaultTab(defaultTabId);
         verify(selectListener).onTabSelected(defaultTabId);
@@ -121,37 +113,37 @@ public class ThreeTabsActivityTest {
     @Test
     @UiThreadTest
     public void afterConfigurationChanged_UserSelectedTabPersistsWhenResettingDefaultTab() {
-        int defaultTabId = R.id.tab_friends;
+        int defaultTabId = com.roughike.bottombar.test.R.id.tab_friends;
 
         bottomBar.setDefaultTab(defaultTabId);
-        bottomBar.selectTabWithId(R.id.tab_nearby);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_nearby);
 
         Bundle savedState = bottomBar.saveState();
         bottomBar.restoreState(savedState);
         bottomBar.setDefaultTab(defaultTabId);
 
         assertNotSame(defaultTabId, bottomBar.getCurrentTabId());
-        assertEquals(R.id.tab_nearby, bottomBar.getCurrentTabId());
+        assertEquals(com.roughike.bottombar.test.R.id.tab_nearby, bottomBar.getCurrentTabId());
     }
 
     @Test
     @UiThreadTest
     public void whenGettingCurrentTab_ReturnsCorrectOne() {
-        int firstTabId = R.id.tab_favorites;
+        int firstTabId = com.roughike.bottombar.test.R.id.tab_favorites;
         bottomBar.selectTabWithId(firstTabId);
 
         assertEquals(firstTabId, bottomBar.getCurrentTabId());
         assertEquals(bottomBar.findPositionForTabWithId(firstTabId), bottomBar.getCurrentTabPosition());
         assertEquals(bottomBar.getTabWithId(firstTabId), bottomBar.getCurrentTab());
 
-        int secondTabId = R.id.tab_nearby;
+        int secondTabId = com.roughike.bottombar.test.R.id.tab_nearby;
         bottomBar.selectTabWithId(secondTabId);
 
         assertEquals(secondTabId, bottomBar.getCurrentTabId());
         assertEquals(bottomBar.findPositionForTabWithId(secondTabId), bottomBar.getCurrentTabPosition());
         assertEquals(bottomBar.getTabWithId(secondTabId), bottomBar.getCurrentTab());
 
-        int thirdTabId = R.id.tab_friends;
+        int thirdTabId = com.roughike.bottombar.test.R.id.tab_friends;
         bottomBar.selectTabWithId(thirdTabId);
 
         assertEquals(thirdTabId, bottomBar.getCurrentTabId());
@@ -165,9 +157,9 @@ public class ThreeTabsActivityTest {
         bottomBar.setOnTabSelectListener(null);
         bottomBar.setOnTabReselectListener(null);
 
-        int firstTabId = R.id.tab_favorites;
-        int secondTabId = R.id.tab_nearby;
-        int thirdTabId = R.id.tab_friends;
+        int firstTabId = com.roughike.bottombar.test.R.id.tab_favorites;
+        int secondTabId = com.roughike.bottombar.test.R.id.tab_nearby;
+        int thirdTabId = com.roughike.bottombar.test.R.id.tab_friends;
 
         bottomBar.selectTabWithId(secondTabId);
         assertOnlyHasOnlyOneSelectedTabWithId(secondTabId);
@@ -195,9 +187,9 @@ public class ThreeTabsActivityTest {
     @UiThreadTest
     public void whenTabIsSelectedOnce_AndNoSelectionListenerSet_ReselectionListenerIsNotFired() {
         bottomBar.setOnTabSelectListener(null);
-        bottomBar.selectTabWithId(R.id.tab_friends);
-        bottomBar.selectTabWithId(R.id.tab_nearby);
-        bottomBar.selectTabWithId(R.id.tab_favorites);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_friends);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_nearby);
+        bottomBar.selectTabWithId(com.roughike.bottombar.test.R.id.tab_favorites);
 
         verifyZeroInteractions(reselectListener);
     }